+++ src/corrigo/__init__.py
@@ -19,7 +19,7 @@ from corrigo.exceptions import (
     ServerError,
 )
 
-__version__ = "0.2.0"
+__version__ = "0.4.0"
 __all__ = [
     "CorrigoClient",
     "CorrigoAuth",